requirements.txt
Django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-environ>=0.11
psycopg[binary]>=3.1
celery[redis]>=5.3
redis>=5.0
Pillow>=10.0
pytest-django>=4.5
gunicorn>=21.0
whitenoise>=6.6
